+++ eslint.config.js
@@ -5,5 +5,6 @@ require('sucrase/register')
 const { defineConfig } = require('./packages/eslint/index.ts')
 
 export default defineConfig([], {
+  gitignore: false,
   sortObjects: true,
 })