+++ packages/eslint/configs/prettier.ts
@@ -1,19 +1,12 @@
 // @ts-expect-error missing types
 import prettierConfig from '@minko-fe/prettier-config'
 import { type FlatESLintConfigItem } from 'eslint-define-config'
-import { configPrettier, pluginPrettier } from '../plugins'
-
-const prettierConflictRules = { ...configPrettier.rules }
-delete prettierConflictRules['vue/html-self-closing']
+import pluginPrettier from 'eslint-plugin-prettier/recommended'
 
 export const prettier: FlatESLintConfigItem[] = [
   {
-    plugins: {
-      prettier: pluginPrettier,
-    },
+    ...pluginPrettier,
     rules: {
-      ...prettierConflictRules,
-      ...pluginPrettier.configs.recommended.rules,
       'prettier/prettier': ['warn', prettierConfig],
     },
   },