+++ packages/eslint/plugins.ts
@@ -1,69 +1,53 @@
-import { interopDefault } from './utils'
-
-/* eslint-disable import/first */
+/* eslint-disable perfectionist/sort-exports */
+/* eslint-disable import/first  */
+// @ts-nocheck
 
 import * as _pluginAntfu from 'eslint-plugin-antfu'
-export const pluginAntfu = interopDefault(_pluginAntfu)
+import { interopDefault } from './utils'
 
-// @ts-expect-error missing types
+export const pluginAntfu: typeof import('eslint-plugin-antfu').default = interopDefault(_pluginAntfu)
+
 import * as _pluginComments from 'eslint-plugin-eslint-comments'
 export const pluginComments = interopDefault(_pluginComments)
 
-// @ts-expect-error missing types
 import * as _pluginMarkdown from 'eslint-plugin-markdown'
 export const pluginMarkdown = interopDefault(_pluginMarkdown)
 
 import * as _pluginTypeScript from '@typescript-eslint/eslint-plugin'
 export const pluginTypeScript = interopDefault(_pluginTypeScript)
 
-// @ts-expect-error missing types
 import * as _pluginReactRefresh from 'eslint-plugin-react-refresh'
 export const pluginReactRefresh = interopDefault(_pluginReactRefresh)
 
-// @ts-expect-error missing types
 import * as _pluginUnicorn from 'eslint-plugin-unicorn'
 export const pluginUnicorn = interopDefault(_pluginUnicorn)
 
-// @ts-expect-error missing types
 import * as _pluginVue from 'eslint-plugin-vue'
 export const pluginVue = interopDefault(_pluginVue)
 
-// @ts-expect-error missing types
 import * as _pluginReact from 'eslint-plugin-react'
 export const pluginReact = interopDefault(_pluginReact)
 
-// @ts-expect-error missing types
 import * as _pluginReactHooks from 'eslint-plugin-react-hooks'
 export const pluginReactHooks = interopDefault(_pluginReactHooks)
 
-// @ts-expect-error missing types
 import * as _pluginDisableAutofix from 'eslint-plugin-disable-autofix'
 export const pluginDisableAutofix = interopDefault(_pluginDisableAutofix)
 
-import * as _pluginPrettier from 'eslint-plugin-prettier'
-export const pluginPrettier = interopDefault(_pluginPrettier)
-
-// @ts-expect-error missing types
-import * as _configPrettier from 'eslint-config-prettier'
-
-export const configPrettier = interopDefault(_configPrettier)
-
-// @ts-expect-error missing types
 export * as parserTypeScript from '@typescript-eslint/parser'
 
-// @ts-expect-error missing types
 export * as pluginImport from 'eslint-plugin-import'
 export * as pluginJsonc from 'eslint-plugin-jsonc'
 
-// @ts-expect-error missing types
 export * as pluginPerfectionist from 'eslint-plugin-perfectionist'
 
-// @ts-expect-error missing types
 export * as pluginUnusedImports from 'eslint-plugin-unused-imports'
 
-// @ts-expect-error missing types
 export * as pluginYml from 'eslint-plugin-yml'
 export * as parserJsonc from 'jsonc-eslint-parser'
 
 export * as parserVue from 'vue-eslint-parser'
 export * as parserYml from 'yaml-eslint-parser'
+
+import * as _pluginNode from 'eslint-plugin-n'
+export const pluginNode = interopDefault(_pluginNode)