+++ packages/eslint/presets.ts
@@ -2,14 +2,18 @@ import { type FlatESLintConfigItem } from 'eslint-define-config'
 import { isPackageExists } from 'local-pkg'
 import {
   _exports,
+  asyncSvelte,
   comments,
+  gitignores,
   ignores,
   imports,
   javascript,
   jsonc,
   markdown,
+  node,
   prettier,
   react,
+  resolveAstro,
   sortObjects,
   sortPackageJson,
   sortTsconfig,
@@ -20,7 +24,7 @@ import {
   yml,
 } from './configs'
 
-export const presetJavaScript = [...javascript, ...comments, ...imports, ..._exports, ...unicorn]
+export const presetJavaScript = [...javascript, ...comments, ...imports, ..._exports, ...unicorn, ...node, ...ignores]
 
 export const presetLangsExtensions = [...markdown, ...yml, ...jsonc, ...sortPackageJson]
 
@@ -34,20 +38,28 @@ const VuePackages = ['vue', 'nuxt', 'vitepress']
 
 const ReactPackages = ['react', 'next']
 
-export function defineConfig(
+const SveltePackages = ['svelte', 'svelte-kit']
+
+const AstroPackages = ['astro']
+
+export async function defineConfig(
   config: FlatESLintConfigItem | FlatESLintConfigItem[] = [],
   {
+    astro: enableAstro = AstroPackages.some((i) => isPackageExists(i)),
     gitignore: enableGitignore = true,
     markdown: enableMarkdown = true,
     prettier: enablePrettier = true,
     react: enableReact = ReactPackages.some((i) => isPackageExists(i)),
     sortObjects: enableSortObjects = false,
+    svelte: enableSvelte = SveltePackages.some((i) => isPackageExists(i)),
     tailwindcss: enableTailwindcss = isPackageExists('tailwindcss'),
     typescript: enableTypescript = isPackageExists('typescript'),
     vue: enableVue = VuePackages.some((i) => isPackageExists(i)),
   }: Partial<{
     react: boolean
     vue: boolean
+    astro: boolean
+    svelte: boolean
     typescript: boolean
     sortObjects: boolean
     prettier: boolean
@@ -55,12 +67,12 @@ export function defineConfig(
     tailwindcss: boolean
     gitignore: boolean
   }> = {},
-): FlatESLintConfigItem[] {
+): Promise<FlatESLintConfigItem[]> {
   const configs = []
 
   configs.push(...presetBasic)
   if (enableGitignore) {
-    configs.push(...ignores)
+    configs.push(...gitignores)
   }
   if (enableReact) {
     configs.push(...react)
@@ -83,6 +95,13 @@ export function defineConfig(
   if (enablePrettier) {
     configs.push(...prettier)
   }
+  if (enableAstro) {
+    configs.push(...resolveAstro(enableTypescript))
+  }
+  if (enableSvelte) {
+    configs.push(...(await asyncSvelte()))
+  }
+
   if (Object.keys(config).length > 0) {
     configs.push(...(Array.isArray(config) ? config : [config]))
   }