+++ packages/eslint/utils/index.ts
@@ -1,3 +1,6 @@
-export function interopDefault(m: any) {
-  return m.default || m
+export type InteropDefault<T> = T extends { default: infer U } ? U : T
+
+/* #__NO_SIDE_EFFECTS__ */
+export function interopDefault<T>(m: T): InteropDefault<T> {
+  return (m as any).default || m
 }