+++ packages/prettier/astro.js
@@ -1,0 +1,12 @@
+/** @type {import("prettier").Config} */
+export default {
+  overrides: [
+    {
+      files: '*.astro',
+      options: {
+        parser: 'astro',
+      },
+    },
+  ],
+  plugins: ['prettier-plugin-astro'],
+}