+++ packages/prettier/svelte.js
@@ -1,0 +1,12 @@
+/** @type {import("prettier").Config} */
+export default {
+  overrides: [
+    {
+      files: '*.svelte',
+      options: {
+        parser: 'svelte',
+      },
+    },
+  ],
+  plugins: ['prettier-plugin-svelte'],
+}