+++ tsup.config.ts
@@ -4,5 +4,7 @@ export default defineConfig({
   clean: true,
   dts: true,
   format: ['esm', 'cjs'],
+  splitting: true,
   target: 'node16',
+  treeshake: true,
 })